improve source code due to 'mvn site' results

+++ realtime/src/main/java/cn/leancloud/im/v2/messages/AVIMFileMessage.java
@@ -153,7 +153,7 @@ public class AVIMFileMessage extends AVIMTypedMessage {
   public long getSize() {
     Map<String, Object> meta = getFileMetaData();
     if (meta != null && meta.containsKey(FILE_SIZE)) {
-      return Long.valueOf(meta.get(FILE_SIZE).toString());
+      return Long.valueOf(meta.get(FILE_SIZE).toString()).longValue();
     }
     return 0;
   }
